I moved the flag images to the `public/assets/flags` directory and `freestyle-shared.css` to the `src` directory. Then, I updated all the paths to correctly reference the flag images and the CSS file.
