+++ src/components/StudyMode/DisplayBlocks/FillGapsBoxBlock.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import { useI18n } from '../../../i18n/I18nContext';
 import './FillGapsBoxBlock.css';
-import '../../Freestyle/freestyle-shared.css';
+import '../../../freestyle-shared.css';
 
 // Helper to shuffle an array
 const shuffleArray = (array) => {
